requirements.txt
openai>=1.14.0
pydantic>=2.6.4
pydantic-core>=2.16.3
python-dotenv>=1.0.0
pytest